By generating an artificial neural network (creating the class), the given flower classification problem has been solved.
It is known that a flower we see in nature is the lily (or iris) plant. It is required to write an algorithm that finds out which of the three different types it belongs to. We have data obtained by measuring over 150 samples from each flower type. For each sample, 4 features (septal length, sepal width, petal length, petal width) and information on which class (species) they belong to are given. Dataset: https://archive.ics.uci.edu/ml/machinelearningdatabases/iris/iris.data and will be used in the project.